refactor: move UserStatus and NotAuthorized in containers
