+++ packages/client/src/app/components/UserSettings.tsx
@@ -4,7 +4,7 @@ import DialogTitle from '@mui/material/DialogTitle';
 import Button from '@mui/material/Button';
 import { XIcon } from '@heroicons/react/outline';
 import { useAuthContext } from '../hooks/useAuthContext';
-import { UserStatus } from './UserStatus';
+import { UserStatus } from '../containers/UserStatus';
 import { StatusIcon } from './StatusIcon';
 
 interface Props {
